server/feat/store-example-2 : 참고용 커밋
